+++ bin/odl2_import_monitor
@@ -5,7 +5,7 @@ OPDS 2.x + ODL collections."""
 
 from webpub_manifest_parser.odl import ODLFeedParserFactory
 
-from palace.manager.api.odl2 import ODL2Importer, ODL2ImportMonitor
+from palace.manager.api.odl2.importer import ODL2Importer, ODL2ImportMonitor
 from palace.manager.core.opds2_import import RWPMManifestParser
 from palace.manager.scripts.monitor import RunCollectionMonitorScript
