--- bin/odl_import_monitor
@@ -1,12 +1,0 @@
-#!/usr/bin/env python
-"""Update the circulation manager server with new books from
-ODL collections."""
-
-
-from palace.manager.api.odl import ODLImporter, ODLImportMonitor
-from palace.manager.scripts.monitor import RunCollectionMonitorScript
-
-RunCollectionMonitorScript(
-    ODLImportMonitor,
-    import_class=ODLImporter,
-).run()
