+++ tests/conftest.py
@@ -7,7 +7,6 @@ pytest_plugins = [
     "tests.fixtures.announcements",
     "tests.fixtures.api_admin",
     "tests.fixtures.api_controller",
-    "tests.fixtures.api_odl",
     "tests.fixtures.api_routes",
     "tests.fixtures.authenticator",
     "tests.fixtures.celery",
@@ -15,7 +14,7 @@ pytest_plugins = [
     "tests.fixtures.files",
     "tests.fixtures.flask",
     "tests.fixtures.library",
-    "tests.fixtures.odl",
+    "tests.fixtures.odl2",
     "tests.fixtures.redis",
     "tests.fixtures.s3",
     "tests.fixtures.search",
