--- tests/files/odl/shared_collection_loan_info.opds
@@ -1,1 +1,0 @@
-<entry xmlns:app="http://www.w3.org/2007/app" xmlns:bib="http://bib.schema.org/" xmlns:bibframe="http://bibframe.org/vocab/" xmlns:dcterms="http://purl.org/dc/terms/" xmlns:drm="http://librarysimplified.org/terms/drm" xmlns:opds="http://opds-spec.org/2010/catalog" xmlns:opf="http://www.idpf.org/2007/opf" xmlns:schema="http://schema.org/" xmlns:simplified="http://librarysimplified.org/terms/" xmlns="http://www.w3.org/2005/Atom" schema:additionalType="http://schema.org/EBook"><id>http://www.feedbooks.com/item/2532629</id><title>How Not to be a Bride</title><author><name>Portia MacIntosh</name><link href="http://localhost:6500/AL/works/contributor/Portia%20MacIntosh/eng/Adult%2CAdults+Only%2CChildren%2CYoung+Adult" type="application/atom+xml;profile=opds-catalog;kind=acquisition" rel="contributor" title="Portia MacIntosh"/></author><summary type="html">&#8216;Delightfully romantic, light-hearted and charmingly entertaining.&#8217; What&#8217;s Better Than Books?\n        Definitely, maybe&#8230;yes?\n        Mia Valentina gave up her high-flying life in LA to move back to Kent over four years ago. But it turns o...</summary><simplified:pwid>2cafaf0f-795c-bb13-b681-16698ecb4cb0</simplified:pwid><link href="http://covers.feedbooks.net/item/2532629.jpg?size=large&amp;t=1508548831" type="image/png" rel="http://opds-spec.org/image"/><link href="http://covers.feedbooks.net/item/2532629.jpg?t=1508548831" type="image/png" rel="http://opds-spec.org/image/thumbnail"/><category term="Adult" scheme="http://schema.org/audience" label="Adult"/><category term="http://librarysimplified.org/terms/fiction/Fiction" scheme="http://librarysimplified.org/terms/fiction/" label="Fiction"/><category term="http://librarysimplified.org/terms/genres/Simplified/Contemporary%20Romance" scheme="http://librarysimplified.org/terms/genres/Simplified/" label="Contemporary Romance"/><category term="http://librarysimplified.org/terms/genres/Simplified/Women%27s%20Fiction" scheme="http://librarysimplified.org/terms/genres/Simplified/" label="Women\'s Fiction"/><category term="http://librarysimplified.org/terms/genres/Simplified/Humorous%20Fiction" scheme="http://librarysimplified.org/terms/genres/Simplified/" label="Humorous Fiction"/><category term="http://librarysimplified.org/terms/genres/Simplified/Literary%20Fiction" scheme="http://librarysimplified.org/terms/genres/Simplified/" label="Literary Fiction"/><dcterms:language>en</dcterms:language><dcterms:publisher>HQ Digital</dcterms:publisher><bibframe:distribution bibframe:ProviderName="DPLA Exchange"/><updated>2018-02-20T03:35:21Z</updated><link href="http://localhost:6500/AL/works/URI/http://www.feedbooks.com/item/2532629" type="application/atom+xml;type=entry;profile=opds-catalog" rel="alternate"/><link href="http://localhost:6500/AL/works/URI/http://www.feedbooks.com/item/2532629/report" rel="issues"/><link href="http://localhost:6500/AL/collections/DPLA%20Exchange/loans/33/fulfill/2" rel="http://opds-spec.org/acquisition" type="application/vnd.adobe.adept+xml"><opds:indirectAcquisition type="application/epub+zip"/><opds:availability since="2018-03-08T17:44:12Z" status="available" until="2018-03-29T17:44:11Z"/><opds:holds total="1"/><opds:copies available="0" total="3"/></link><link href="http://localhost:6500/AL/collections/DPLA%20Exchange/loans/33/revoke" rel="http://librarysimplified.org/terms/rel/revoke"/><link href="http://localhost:6500/AL/collections/DPLA%20Exchange/loans/33" rel="self"/><link href="http://localhost:6500/AL/works/URI/http://www.feedbooks.com/item/2532629/related_books" type="application/atom+xml;profile=opds-catalog;kind=acquisition" rel="related" title="Recommended Works"/><link href="http://localhost:6500/AL/annotations/URI/http://www.feedbooks.com/item/2532629/" type="application/ld+json; profile=&quot;http://www.w3.org/ns/anno.jsonld&quot;" rel="http://www.w3.org/ns/oa#annotationService"/></entry>
